+++ gulpfile.js
@@ -4,7 +4,7 @@ const gulpClean = require('gulp-clean')
 const sass = require('sass')
 const yupSass = require('gulp-sass')(sass)
 const gulpZip = require('gulp-zip')
-const { compile } = require('nexe')
+const pkg = require('pkg')
 
 const chromeManifest = require('./src/browser-ext/vendor/chrome/manifest.json')
 const firefoxManifest = require('./src/browser-ext/vendor/firefox/manifest.json')
@@ -12,19 +12,23 @@ const operaManifest = require('./src/browser-ext/vendor/opera/manifest.json')
 const edgeManifest = require('./src/browser-ext/vendor/edge/manifest.json')
 const pcPackageJSON = require('./src/pc/package.json')
 
-const FILE_NAME = {
-  CHROME: `ZaDark-Chrome-${chromeManifest.version}`,
-  FIREFOX: `ZaDark-Firefox-${firefoxManifest.version}`,
-  OPERA: `ZaDark-Opera-${operaManifest.version}`,
-  EDGE: `ZaDark-Edge-${edgeManifest.version}`,
-  MACOS: `ZaDark-macOS-${pcPackageJSON.version}`,
-  WINDOWS: `ZaDark-Windows-${pcPackageJSON.version}`
+const dot2Underscore = (v = '') => v.replace(/\./g, '_')
+
+const DIST_FILE_NAME = {
+  CHROME: `ZaDark-Chrome-${dot2Underscore(chromeManifest.version)}`,
+  FIREFOX: `ZaDark-Firefox-${dot2Underscore(firefoxManifest.version)}`,
+  OPERA: `ZaDark-Opera-${dot2Underscore(operaManifest.version)}`,
+  EDGE: `ZaDark-Edge-${dot2Underscore(edgeManifest.version)}`,
+  MACOS: `ZaDark-macOS-${dot2Underscore(pcPackageJSON.version)}`,
+  WINDOWS: `ZaDark-Windows-${dot2Underscore(pcPackageJSON.version)}`
 }
 
 // Clean
+
 const cleanBuild = () => {
   return src('build', { read: false, allowEmpty: true }).pipe(gulpClean())
 }
+
 const cleanDist = () => {
   return src('dist', { read: false, allowEmpty: true }).pipe(gulpClean())
 }
@@ -56,8 +60,6 @@ const buildBrowserExt = (browser) => {
     src(`./src/browser-ext/vendor/${browser}/browser.js`).pipe(dest(`./build/${browser}/js`)),
     src(`./src/browser-ext/vendor/${browser}/background.js`).pipe(dest(`./build/${browser}/js`)),
 
-    src(`./src/browser-ext/vendor/${browser}/welcome.html`).pipe(dest(`./build/${browser}`)),
-
     src('./src/browser-ext/libs/**/*').pipe(dest(`./build/${browser}/libs`)),
     src('./src/browser-ext/js/**/*').pipe(dest(`./build/${browser}/js`)),
     src('./src/browser-ext/css/**/*').pipe(dest(`./build/${browser}/css`)),
@@ -87,18 +89,16 @@ const buildPC = () => {
   return src('./src/pc/**/*').pipe(dest('./build/pc'))
 }
 
-// Compile
+// Package
 
-const compileMacOS = () => {
+const pkgMacOS = () => {
   return new Promise((resolve, reject) => {
-    // macOS
-    compile({
-      input: './build/pc/index.js',
-      output: `./dist/macOS/${FILE_NAME.MACOS}.command`,
-      resources: './build/pc/assets/**/*',
-      targets: 'mac-x64-14.15.3',
-      silent: true
-    }).then(() => {
+    pkg.exec([
+      './build/pc/index.js',
+      '--config', './pkg.config.json',
+      '--targets', 'node14-macos-x64',
+      '--output', `./dist/macos/${DIST_FILE_NAME.MACOS}`
+    ]).then(() => {
       resolve(true)
     }).catch((error) => {
       reject(error)
@@ -106,16 +106,14 @@ const compileMacOS = () => {
   })
 }
 
-const compileWindows = () => {
+const pkgWindows = () => {
   return new Promise((resolve, reject) => {
-    // Windows
-    compile({
-      input: './build/pc/index.js',
-      output: `./dist/Windows/${FILE_NAME.WINDOWS}.exe`,
-      resources: './build/pc/assets/**/*',
-      targets: 'windows-x86-14.15.3',
-      silent: true
-    }).then(() => {
+    pkg.exec([
+      './build/pc/index.js',
+      '--config', './pkg.config.json',
+      '--targets', 'node14-win-x64',
+      '--output', `./dist/windows/${DIST_FILE_NAME.WINDOWS}`
+    ]).then(() => {
       resolve(true)
     }).catch((error) => {
       reject(error)
@@ -124,14 +122,16 @@ const compileWindows = () => {
 }
 
 // Zip
+
 const zipMacOS = () => {
-  return src(`./dist/macOS/${FILE_NAME.MACOS}.command`)
-    .pipe(gulpZip(`${FILE_NAME.MACOS}.zip`))
+  return src(`./dist/macOS/${DIST_FILE_NAME.MACOS}`)
+    .pipe(gulpZip(`${DIST_FILE_NAME.MACOS}.zip`))
     .pipe(dest('./dist/macOS'))
 }
+
 const zipWindows = () => {
-  return src(`./dist/Windows/${FILE_NAME.WINDOWS}.exe`)
-    .pipe(gulpZip(`${FILE_NAME.WINDOWS}.zip`))
+  return src(`./dist/Windows/${DIST_FILE_NAME.WINDOWS}.exe`)
+    .pipe(gulpZip(`${DIST_FILE_NAME.WINDOWS}.zip`))
     .pipe(dest('./dist/Windows'))
 }
 
@@ -139,30 +139,33 @@ const zipWindows = () => {
 
 const chromeDist = () => {
   return src('./build/chrome/**')
-    .pipe(gulpZip(`${FILE_NAME.CHROME}.zip`))
+    .pipe(gulpZip(`${DIST_FILE_NAME.CHROME}.zip`))
     .pipe(dest('./dist/chrome'))
 }
 
 const firefoxDist = () => {
   return src('./build/firefox/**')
-    .pipe(gulpZip(`${FILE_NAME.FIREFOX}.zip`))
+    .pipe(gulpZip(`${DIST_FILE_NAME.FIREFOX}.zip`))
     .pipe(dest('./dist/firefox'))
 }
 
 const operaDist = () => {
   return src('./build/opera/**')
-    .pipe(gulpZip(`${FILE_NAME.OPERA}.zip`))
+    .pipe(gulpZip(`${DIST_FILE_NAME.OPERA}.zip`))
     .pipe(dest('./dist/opera'))
 }
 
 const edgeDist = () => {
   return src('./build/edge/**')
-    .pipe(gulpZip(`${FILE_NAME.EDGE}.zip`))
+    .pipe(gulpZip(`${DIST_FILE_NAME.EDGE}.zip`))
     .pipe(dest('./dist/edge'))
 }
 
-const macOSDist = series(compileMacOS, zipMacOS)
-const windowsDist = series(compileWindows, zipWindows)
+const pcDist = series(
+  pkgMacOS,
+  pkgWindows,
+  parallel(zipMacOS, zipWindows)
+)
 
 // Exports
 
@@ -187,8 +190,7 @@ const distAll = series(
     firefoxDist,
     operaDist,
     edgeDist,
-    macOSDist,
-    windowsDist
+    pcDist
   )
 )